requirements.txt
ultralytics==8.4.37
opencv-python-headless==4.13.0.92
streamlit==1.56.0
reportlab==4.4.1
Pillow==12.2.0
numpy==2.4.4